FunctionProject/requirements.txt
azure-functions
Pillow==6.2.1
keras==2.2.4
tensorflow==1.13.1